refactor: API 변경 내용 적용
