--- frontend/src/apis/petPlants.ts
@@ -1,11 +1,0 @@
-import { NewPetPlantRequest } from 'types/api/petPlant';
-
-const postForm = (form: NewPetPlantRequest) => {
-  return fetch('/pet-plants', { method: 'POST', body: JSON.stringify({ ...form }) });
-};
-
-const petPlantsAPI = {
-  postForm,
-};
-
-export default petPlantsAPI;
